+++ limen/sfd/reference_architecture/random_binary.py
@@ -77,7 +77,8 @@ class RandomBinary(ReferenceModel):
 
         if inline_metrics:
             results.update(self._compute_confusion(preds, data['y_test']))
-            results.update(self._compute_backtest(preds, data, actuals=data['y_test']))
+            results.update(self._compute_confusion_return_metrics(preds, data['y_test'], data))
+            results.update(self._compute_backtest(preds, data))
 
         return results